# src/openapi.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AlexMubarakshin/ApiSchema/1.0.1
info:
  description: Api schema
  version: "1.0.1"
  title: API schema
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/account/register:
    post:
      tags:
        - account
      description: "Регистрация пользователя"
      summary: "Регистрация"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "E-Mail"
                  example: "arkady@example.com"
                passwd1:
                  type: string
                  description: "Пароль"
                  example: "123456"
                passwd2:
                  type: string
                  description: "Повтор пароля"
                  example: "123456"
                firstName:
                  type: string
                  description: "Имя"
                  example: "Аркадий"
                lastName:
                  type: string
                  description: "Фамилия"
                  example: "Паровозов"
                code:
                  type: string
                  description: "Код"
                  example: "3675"
      responses:
        "200":
          description: "Успешная регистрация"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "ok"
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/account/password/recovery:
    put:
      tags:
        - account
      description: "Восстановить пароль по email.<br>Высылает ссылку для восстановления на почтовый ящик."
      summary: "Восстановление пароля"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "E-Mail"
                  example: "alex@example.com"
      responses:
        "200":
          description: "Успешный запрос"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "ok"
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/account/password/set:
    put:
      tags:
        - account
      description: "Изменить пароль."
      summary: "Смена пароля"
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: "Токен высланый на почтоый ящик. (Параметр используется в случае если пользователь восстанавливает пароль по почте)"
                  example: "ad1fj118f11"
                password:
                  type: string
                  description: "Новый пароль"
                  example: "123456"
      responses:
        "200":
          description: "Успешная смена пароля"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "ok"
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/account/email:
    put:
      tags:
        - account
      description: "Смена email'а пользователя."
      summary: "Смена email'а"
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "Новый email"
                  example: "arkady-ukupnik@thebest.com"
      responses:
        "200":
          description: "Успешная смена email'a"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "ok"
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/account/notifications:
    put:
      tags:
        - account
      description: "Редактирование настроек оповещения"
      summary: "Редактирование настроек оповещения"
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notification_type:
                  $ref: "#/components/schemas/NotificationTypeCode"
                enable:
                  type: boolean
                  description: "Уведомлять о событиях данного типа"
                  example: true
      responses:
        "200":
          description: "Успешная смена статуса"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "ok"
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/notifications:
    get:
      tags:
        - notifications
      description: "Получить список уведомлений для текущего пользователя"
      summary: "Получить список уведомлений"
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: "Успешное получение списка уведомлений"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/NotificationBase'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - notifications
      description: "Отметить уведомление как \"прочитанное\""
      summary: "Прочитать уведомление"
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          description: "ID уведомления"
          required: false
          schema:
            type: number
            example: 1
        - name: all
          in: query
          description: "Пометка что прочесть все уведомления"
          required: false
          schema:
            type: boolean
            example: true

      responses:
        "200":
          description: "Успешное получение списка уведомлений"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/NotificationBase'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Тип ошибки"
          example: "invalid_request"
          enum:
            - invalid_request
            - unauthorized
        error_description:
          type: string
          items:
            $ref: '#/components/schemas/ErrorDescription'
    ErrorDescription:
      type: object
      properties:
        error_description:
          type: string
          description: "Error description"
          enum:
            - "Unauthorized"
            - "An authorization code must be supplied"
            - "Redirect URI mismatch"
            - "Invalid authorization code: CODE"
            - "Bad credentials"
    Response:
      type: object
      properties:
        error:
          type: string
          description: "Сообщение, в случае ошибки здесь будет причина"
        timestamp:
          type: number
          example: 1559751301818
    ListResponse:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            total:
              type: integer
              description: "Количество элементов в списке"
              format: int64
            offset:
              type: integer
              description: "Отступ от начала списка"
              format: int64
            perPage:
              type: integer
              description: "Количество элементов на страницу"
              format: int64
              default: 20
...
    NotificationBase:
      type: object
      description: "Оповещения"
      properties:
        id:
          type: number
          example: 1
        type_id:
          type: number
          example: 1
          description: "ID типа оповещения"
        sent_time:
          type: number
          example: 1559751301818
          description: "Время отправки"
        entity_id:
          type: number
          example: 1
          description: "Идентификатор сущности, относительно которой отправлено оповещение (комментарий, друг, пост или сообщение)"
        info:
          type: string
          description: "Текст уведомления"
...
    NotificationTypeCode:
      type: string
      description: "Код типа уведомлений"
      enum:
        - POST
        - POST_COMMENT
        - COMMENT_COMMENT
        - FRIEND_REQUEST
        - MESSAGE
...
  parameters:
    offsetParam:
      in: query
      description: "Отступ от начала списка"
      name: offset
      required: false
      schema:
        type: integer
        format: int64
    limitParam:
      in: query
      description: "Количество элементов на страницу"
      name: itemPerPage
      required: false
      schema:
        default: 20
        type: integer
        format: int64